requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
dj-database-url==1.2.0
Django==4.1.7
django-bootstrap==0.2.4
django-bootstrap-v5==1.0.11
django-resized==1.0.2
django-storages==1.13.2
gunicorn==20.1.0
Pillow==9.4.0
psycopg2==2.9.5
soupsieve==2.4
sqlparse==0.4.3
